+++ MergeSorts/mergesort.h
@@ -1,5 +1,6 @@
 #include <stdint.h>
+#include <stddef.h>
 
-void mergeSortNaive(int32_t* array, uint32_t size);
-void mergeSort2Array(int32_t* array, uint32_t size);
-void mergeSortMultiThreaded(int32_t* array, uint32_t size);
+void mergeSortNaive(int32_t* array, size_t size);
+void mergeSort2Array(int32_t* array, size_t size);
+void mergeSortMultiThreaded(int32_t* array, size_t size);
